The "Ethernet offset" is 16 bits in the file; make it so in the pseudo-header.
This should squelch warnings from Ida7b98af8c44a52ddac2c4ab0702db2519a0c4af.

Reviewed-on: https://code.wireshark.org/review/28573

+++ wiretap/wtap.h
@@ -1112,7 +1112,7 @@ struct nstr_phdr {
     guint8 nicno_len;
     guint8 dir_offset;
     guint8 dir_len;
-    guint8 eth_offset;
+    guint16 eth_offset;
     guint8 pcb_offset;
     guint8 l_pcb_offset;
     guint8 rec_type;
